requirements.txt.txt
Flask
numpy
opencv-python
tensorflow
Pillow